+++ src/main/java/com/jt/customer/entity/CustomerInterface.java
@@ -3,5 +3,5 @@ package com.jt.customer.entity;
 public interface CustomerInterface {
   String getDate();
   int getRole();
-  long getTotal();
+  int getTotal();
 }
